DrawMapをクラス分けした
