--- procon29-kyogi/Siv3D(August2016v2)1/Agent.cpp
@@ -1,9 +1,0 @@
-#include "MapClass/Agent.h"
-
-Agent* Agent::InsAgent = nullptr;
-Agent* Agent::getAgent() {
-	if (InsAgent == nullptr) {
-		InsAgent = new Agent;
-	}
-	return InsAgent;
-}
