+++ procon29-kyogi/Siv3D(August2016v2)1/Algorithm/Prefetching.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../MapClass/Map.h"
+#include "../Data/Map.h"
 #include <vector>
 
 class Prefetching {
