+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Agent.cpp
@@ -1,0 +1,9 @@
+#include "Agent.h"
+
+Agent* Agent::InsAgent = nullptr;
+Agent* Agent::getAgent() {
+	if (InsAgent == nullptr) {
+		InsAgent = new Agent;
+	}
+	return InsAgent;
+}
