+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Evaluaiton.cpp
@@ -1,0 +1,11 @@
+#include "Evaluation.h"
+
+Evaluation* Evaluation::InsEvaluation = nullptr;
+Evaluation* Evaluation::getEvaluation()
+{
+	if (InsEvaluation == nullptr)
+	{
+		InsEvaluation = new Evaluation;
+	}
+	return InsEvaluation;
+}
