+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Evaluation.h
@@ -1,0 +1,16 @@
+#pragma once
+//周辺八近傍
+#define AroundEight 8
+
+class  Evaluation{
+public:
+	static Evaluation* getEvaluation();
+	//合計点による評価点
+	int SumScore[AroundEight];
+	//タイルによる評価点
+	int TileScore[AroundEight];
+	//八近傍の移動可能マス
+	int Movable[AroundEight];
+private:
+	static Evaluation *InsEvaluation;
+};
