+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Map.cpp
@@ -1,0 +1,9 @@
+#include "Map.h"
+
+Map* Map::InsMap = nullptr;
+Map* Map::getMap() {
+	if (InsMap == nullptr) {
+		InsMap = new Map;
+	}
+	return InsMap;
+}
