+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Masu.cpp
@@ -1,0 +1,9 @@
+#include "Masu.h"
+Masu* Masu::InsMasu = nullptr;
+
+Masu* Masu::getMasu() {
+	if (InsMasu == nullptr) {
+		InsMasu = new Masu;
+	}
+	return InsMasu;
+}
