+++ procon29-kyogi/Siv3D(August2016v2)1/Data/Masu.h
@@ -1,0 +1,14 @@
+#pragma once
+class Masu {
+public :
+	static Masu* getMasu();
+	//タイルのポイント
+	int TilePoint;
+	//Non-何もなし,FriendTile-味方のタイル,EnemyTile-敵のタイル
+	//置かれているタイルを管理
+	enum StateOfMasu{ Non = 0, FriendTile = 1, EnemyTile = 2 };
+private:
+	static Masu* InsMasu;
+};
+
+
