+++ procon29-kyogi/Siv3D(August2016v2)1/Data/TurnManager.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "Map.h"
+#define TurnValue 120
+
+class TurnManager {
+public :
+	TurnManager *getTurnManager();
+	Map Turn[TurnValue];
+private:
+	static TurnManager * InsTurnManager;
+};
