--- procon29-kyogi/Siv3D(August2016v2)1/Evaluaiton.cpp
@@ -1,11 +1,0 @@
-#include "MapClass/Evaluation.h"
-
-Evaluation* Evaluation::InsEvaluation = nullptr;
-Evaluation* Evaluation::getEvaluation()
-{
-	if (InsEvaluation == nullptr)
-	{
-		InsEvaluation = new Evaluation;
-	}
-	return InsEvaluation;
-}
