+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/CreateMapClass.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "ReadQR.h"
-#include "../MapClass/Map.h"
+#include "../Data/Map.h"
 
 class CreateMapClass {
 public:
