+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawLeft.cpp
@@ -1,0 +1,26 @@
+#include "DrawLeft.h"
+
+
+void DrawLeft::drawLeftManager(void)
+{
+	//Mapの表示
+}
+
+void DrawLeft::drawMap(void) {
+
+}
+
+void DrawLeft::drawAgent(void)
+{
+	//Mapに重ねてAgent位置の表示(両方)
+}
+
+void DrawLeft::drawTilePoint(void)
+{
+	//TilePointをMapに重ねる形で表示
+}
+
+void DrawLeft::drawMovableTile(void)
+{
+	//自分のエージェントが移動出来る場所の表示
+}
