+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawLeft.h
@@ -1,0 +1,11 @@
+#pragma once
+
+class DrawLeft {
+public:
+	void drawLeftManager(void);
+private:
+	void drawMap(void);
+	void drawAgent(void);
+	void drawTilePoint(void);
+	void drawMovableTile(void);
+};
