+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawMap.cpp
@@ -3,35 +3,4 @@
 void DrawMap::drawManager(void)
 {
 
-	//drawの制御
-}
-
-void DrawMap::drawMap(void)
-{
-	//Mapの表示
-}
-
-void DrawMap::drawAgent(void)
-{
-	//Mapに重ねてAgent位置の表示（両方）
-}
-
-void DrawMap::drawTilePoint(void)
-{
-	//TilePointをMapに重ねる形で表示
-}
-
-void DrawMap::drawScore(void)
-{
-	//両チームの合計得点の表示
-}
-
-void DrawMap::drawMovableTile(void)
-{
-	//自分のエージェントが移動出来る場所の表示
-}
-
-void DrawMap::drawTurn(void)
-{
-	//何ターンか表示
 }
