+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawMap.h
@@ -3,10 +3,4 @@ class DrawMap {
 public:
 	void drawManager(void);
 private:
-	void drawMap(void);
-	void drawAgent(void);
-	void drawTilePoint(void);
-	void drawScore(void);
-	void drawMovableTile(void);
-	void drawTurn(void);
 };
