+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawRight.cpp
@@ -1,0 +1,16 @@
+#include "DrawRight.h"
+
+void DrawLeft::drawRightManager(void)
+{
+	//右側の表示を管理
+}
+
+void DrawLeft::drawTurn(void)
+{
+	//ターン数を表示
+}
+
+void DrawLeft::drawScore(void)
+{
+	//両チームスコアを表示
+}
