+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/DrawRight.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class DrawLeft {
+public :
+	void drawRightManager(void);
+private : 
+	void drawTurn(void);
+	void drawScore(void);
+};
