+++ procon29-kyogi/Siv3D(August2016v2)1/IOManager/Interrupt.h
@@ -1,6 +1,6 @@
 #pragma once
 //#include "../MapClass/Map.h"
-#include "../MapClass/Evaluation.h"
+#include "../Data/Evaluation.h"
 
 class Interrupt {
 public:
