--- procon29-kyogi/Siv3D(August2016v2)1/Map.cpp
@@ -1,9 +1,0 @@
-#include "MapClass/Map.h"
-
-Map* Map::InsMap = nullptr;
-Map* Map::getMap() {
-	if (InsMap == nullptr) {
-		InsMap = new Map;
-	}
-	return InsMap;
-}
