--- procon29-kyogi/Siv3D(August2016v2)1/Masu.cpp
@@ -1,9 +1,0 @@
-#include "MapClass/Masu.h"
-Masu* Masu::InsMasu = nullptr;
-
-Masu* Masu::getMasu() {
-	if (InsMasu == nullptr) {
-		InsMasu = new Masu;
-	}
-	return InsMasu;
-}
