--- procon29-kyogi/Siv3D(August2016v2)1/TurnManager.cpp
@@ -1,9 +1,0 @@
-#include "TurnManager.h"
-
-TurnManager* TurnManager::InsTurnManager = nullptr;
-TurnManager* TurnManager::getTurnManager() {
-	if (InsTurnManager == nullptr) {
-		InsTurnManager = new TurnManager;
-	}
-	return InsTurnManager;
-}
