--- procon29-kyogi/Siv3D(August2016v2)1/TurnManager.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "MapClass/Map.h"
-#define TurnValue 120
-
-class TurnManager {
-public :
-	TurnManager *getTurnManager();
-	Map Turn[TurnValue];
-private:
-	static TurnManager * InsTurnManager;
-};
